Allow reuse of compression/decompression buffers
This commit adds a new member to the program_runtime structure that
instructs the compression and decompression routines to reuse the
buffers previously allocated on the GPU. This is especially important
to provide snappy-cuda as a compression filter for HDF5, as several
calls to the (de)compressor are expected, all of which using the same
fixed chunk size (except for the last block processed).

A control flag that previously existed only in the main application has
also been moved to the program_runtime structure: a boolean that tells
if the GPU is supposed to be used to process the input buffers.

Two new API are introduced to clean up resources:
terminate_compression() and terminate_decompression().