Implement an I/O filter for HDF5
This commit adds an implementation of an I/O filter for HDF5 that
compresses and decompresses data streams using the Snappy CUDA APIs.
Both the CPU and the GPU implementations are supported at this time.

There are two tunnables (via setenv):
- SNAPPY_USE_CUDA=0    disables the use of GPU (enabled by default)
- SNAPPY_BLOCKSIZE=N   compression block size (default: 32k)

Although HDF5 doesn't have an I/O filter API designed to support a
reusable context, we do our best to avoid memory deallocations. Critical
data structures are allocated once and deallocated when the calling
process exits (with an atexit() callback).