read_varint32, read_uint32: take input from non-managed memory
NVIDIA GDS expects to take input buffers explicitly managed by
the user (i.e., it cannot handle cudaMallocManaged buffers).
This commit modifies read_varint32() and read_uint32() so they
explicitly memcopy data from the device to host when the function
is called in host context.

There are more functions that could be modified too, such as
read_long_literal_size() and make_offset_1_byte(). We won't
make changes to those functions for now, as the input data we
have been using to validate this interface are not triggering
their execution.